refactor(agents): bind approval hosts at Gateway ingress
