+++ docs/.generated/plugin-sdk-api-baseline.sha256
@@ -4,7 +4,7 @@ cbf4e2c3088f8886a7c9ea91325a66e0f0846cea21f0b2891f36399b4811306c  module/account
 8e985f345f21a1c9a2b0e94304aaaad6a326bec1c1ce3b26027d2862804a366e  module/account-resolution
 e5e67ddf3cab38fcbf9220bc3160715897e2709d9a9ff6ff36f1ecc9453c2367  module/agent-config-primitives
 74daa746deb548379d3f0d6eac3c4d082df1034c4360cc03bf51fee0f10a2e4d  module/agent-harness
-98c427c987ed9ed54c8ba82c54b9ff0d019d87d7b52066de79aa7f34c449adee  module/agent-harness-runtime
+e8a50f1f52594b5a910a3b284ce95509584dd88e7d1508e6e57fa220d586362f  module/agent-harness-runtime
 5168648cd946abad8a92822889f13ceacc87ed502314a66190d0b1eb8ebe76ea  module/agent-media-payload
 2dcb4d62d90e5d71594f6b843c97534509a154784e378fb1c75bd86b5122b710  module/agent-runtime
 56b6d5fb6af3d95af1200065aca2e7d4f59e5fa59740505fe6ff433077ef6646  module/allow-from
@@ -121,7 +121,7 @@ aa8a411ad37c1d1143b67376bf2d20255b9eedff61d80815f42e4f8ed7bd8e58  module/secret-
 57dcb1462d4c4f9a98d934c4ca975b163d704758af9821a64001ff3ac05637c3  module/secret-input-runtime
 e576b537880f63b3a91f3608f7e84c873bce6c6a3d9a0ba98c247f46de788d25  module/secret-ref-runtime
 62ccaafc8e0677e850339f4a4333f9f16ae9fed979bcef003890b2a47507147f  module/security-runtime
-673c64502fdffb2d6361a7cf2ad0c33ffe15707b5e5027de1d88701ce3d8ade1  module/session-catalog
+66c6fca7a92ab6e5013e5235b18a7e186f987d87016d77db52577c1385073b82  module/session-catalog
 50f5e344f98c27570b7a30e32a906b612e2383d21f102e88cd93e1d5425a6de9  module/session-discussion
 f112bdabc51ba8659b37d0a6f6a32a2b1d471e5b49b56e108bf750ec55a7ea71  module/session-store-runtime
 36affbe151431a6141664b6838e20f2d121ff210d57a3c1b4b41a8818b5c81d8  module/setup
